requirements.txt
tabula
python-dotenv
pysqlcipher3